+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+
+## 31.11.1 - 2018-08-06
 ### Changed
 - [Patch] Allow Node types for SelectDropdown label prop
 - [Patch] Allow numbers for TextField value prop